Add some basic unit tests
