+++ build.gradle
@@ -49,6 +49,7 @@ allprojects {
 
     dependencies {
         testImplementation 'org.junit.jupiter:junit-jupiter-api:5.10.0'
+        testImplementation 'org.junit.jupiter:junit-jupiter-params:5.10.0'
         testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.10.0'
     }
