+++ common/src/test/java/net/william278/husksync/hook/PlanDataExtensionTests.java
@@ -20,19 +20,16 @@
 package net.william278.husksync.hook;
 
 import com.djrapitops.plan.extension.extractor.ExtensionExtractor;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-/**
- * Tests for the {@link PlanHook} and {@link PlanDataExtension} implementation
- */
-public class PlanHookTests {
+@DisplayName("Plan Hook Tests")
+public class PlanDataExtensionTests {
 
-    /**
-     * Throws IllegalArgumentException if there is an implementation error or warning.
-     */
     @Test
-    public void testExtensionImplementationErrors() {
-        new ExtensionExtractor(new PlanDataExtension()).validateAnnotations();
+    @DisplayName("Test Plan Hook Implementation")
+    public void testPlanHookImplementation() {
+        new ExtensionExtractor(new PlanHook.PlanDataExtension()).validateAnnotations();
     }
 
 }
